requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
proxmoxer==1.3.1
requests==2.28.2
Flask-WTF==1.1.1
Flask-Bootstrap==3.3.7.1
python-dotenv==1.0.0